issue-627: merge Issues + Agent floating bubbles into single Help menu
Replaces the two corner FABs (red exclamation IssuesWidget at right:24px
+ green sparkles AgentWidget at right:96px) with one Help bubble that
opens a small popover menu with two items, in this order:

  1. Talk with AI Agent  (primary, listed first)
  2. Create issue        (secondary)

Item 1 opens the agent panel; the existing in-panel consent gate handles
first-time consent. The gate now lists explicit caveats — full
conversation log visible to user AND admins, history retained 90 days
then deleted — surfaced at the right moment instead of buried inside
the Legal doc.

Item 2 opens the existing Issues modal unchanged.

- New: ViewComponents/HelpWidgetViewComponent.cs +
  Views/Shared/Components/HelpWidget/Default.cshtml. Folds in the
  former IssuesWidget's modal (verbatim) and AgentWidget's panel +
  consent gate. Phase 1: agent option only shown to Admins (matches
  former AgentWidget gate); other authenticated users get the help
  bubble with only "Create issue".
- Deleted: ViewComponents/{Agent,Issues}WidgetViewComponent.cs and
  Views/Shared/Components/{Agent,Issues}Widget/.
- _Layout.cshtml + _AdminLayout.cshtml now invoke "HelpWidget".
- widget.js no longer depends on agentWidgetLauncher (the menu opens
  the panel); guards on closeBtn to keep the script reusable if the
  panel is rendered without one.
- Added Help_* resource keys (en, ca, de, es, fr, it) for the menu
  items + caveat copy.
- agent.css: help-widget-menu-item styling with min-height:44px tap
  targets and a mobile clamp (max-width: calc(100vw - 32px)).

Closes nobodies-collective/Humans#627
