+++ src/Humans.Web/Resources/SharedResource.ca.resx
@@ -1831,6 +1831,11 @@
   <data name="Agent_ViewHistory" xml:space="preserve"><value>Historial</value></data>
   <data name="Agent_Send" xml:space="preserve"><value>Envia</value></data>
   <data name="Agent_WidgetButton" xml:space="preserve"><value>Obrir l'assistent</value></data>
+  <data name="Help_WidgetButton" xml:space="preserve"><value>Ajuda</value></data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve"><value>Parlar amb l'agent IA</value></data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve"><value>Crear incidència</value></data>
+  <data name="Help_Agent_Caveat_LogVisibility" xml:space="preserve"><value>El registre complet de la conversa és visible per a tu i per als administradors per a revisió.</value></data>
+  <data name="Help_Agent_Caveat_Retention" xml:space="preserve"><value>L'historial es conserva durant 90 dies i després s'elimina.</value></data>
   <data name="Nav_Agent" xml:space="preserve"><value>Assistent</value></data>
