+++ src/Humans.Web/Resources/SharedResource.de.resx
@@ -1829,6 +1829,11 @@
   <data name="Agent_ViewHistory" xml:space="preserve"><value>Verlauf</value></data>
   <data name="Agent_Send" xml:space="preserve"><value>Senden</value></data>
   <data name="Agent_WidgetButton" xml:space="preserve"><value>Assistent öffnen</value></data>
+  <data name="Help_WidgetButton" xml:space="preserve"><value>Hilfe</value></data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve"><value>Mit KI-Assistent sprechen</value></data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve"><value>Problem melden</value></data>
+  <data name="Help_Agent_Caveat_LogVisibility" xml:space="preserve"><value>Das vollständige Gesprächsprotokoll ist für dich und für Administratoren zur Überprüfung sichtbar.</value></data>
+  <data name="Help_Agent_Caveat_Retention" xml:space="preserve"><value>Der Verlauf wird 90 Tage lang aufbewahrt und dann gelöscht.</value></data>
   <data name="Nav_Agent" xml:space="preserve"><value>Assistent</value></data>
