+++ src/Humans.Web/Resources/SharedResource.fr.resx
@@ -1829,6 +1829,11 @@
   <data name="Agent_ViewHistory" xml:space="preserve"><value>Historique</value></data>
   <data name="Agent_Send" xml:space="preserve"><value>Envoyer</value></data>
   <data name="Agent_WidgetButton" xml:space="preserve"><value>Ouvrir l'assistant</value></data>
+  <data name="Help_WidgetButton" xml:space="preserve"><value>Aide</value></data>
+  <data name="Help_Menu_TalkWithAgent" xml:space="preserve"><value>Parler avec l'agent IA</value></data>
+  <data name="Help_Menu_CreateIssue" xml:space="preserve"><value>Créer un signalement</value></data>
+  <data name="Help_Agent_Caveat_LogVisibility" xml:space="preserve"><value>Le journal complet de la conversation est visible pour toi et pour les administrateurs à des fins de revue.</value></data>
+  <data name="Help_Agent_Caveat_Retention" xml:space="preserve"><value>L'historique est conservé pendant 90 jours, puis supprimé.</value></data>
   <data name="Nav_Agent" xml:space="preserve"><value>Assistant</value></data>
