+++ src/Humans.Web/Views/Shared/_AdminLayout.cshtml
@@ -81,6 +81,6 @@
             crossorigin="anonymous"></script>
     <script src="~/js/site.js" asp-append-version="true"></script>
     @await RenderSectionAsync("Scripts", required: false)
-    @await Component.InvokeAsync("IssuesWidget")
+    @await Component.InvokeAsync("HelpWidget")
 </body>
 </html>
