+++ src/Humans.Web/wwwroot/css/agent.css
@@ -8,3 +8,32 @@
 @media (prefers-reduced-motion: reduce) {
     .agent-panel { transition: none !important; }
 }
+
+/* Help widget — single FAB + popover menu replacing former
+   IssuesWidget + AgentWidget bubbles (issue #627). */
+.help-widget-menu-item {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    padding: 12px 16px;
+    min-height: 44px;
+    background: transparent;
+    border: 0;
+    text-align: left;
+    cursor: pointer;
+    color: inherit;
+}
+.help-widget-menu-item:hover,
+.help-widget-menu-item:focus {
+    background: #f4f4f4;
+    outline: none;
+}
+.help-widget-menu-item-primary {
+    font-weight: 600;
+}
+@media (max-width: 576px) {
+    .help-widget-menu {
+        right: 0 !important;
+        max-width: calc(100vw - 32px);
+    }
+}
